requirements.txt
# GUI依赖
customtkinter>=5.2.2
pillow>=10.2.0
piexif>=1.1.3
imagehash>=4.3.1

# 测试依赖
pytest>=8.0.0
pytest-cov>=4.1.0

tqdm==4.66.2
tkcalendar==1.6.1  # 日历选择控件
numpy==1.26.4  # imagehash的依赖 